% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDesign.R
\name{generateDesign}
\alias{generateDesign}
\title{Generate Response and Design Matrix}
\usage{
generateDesign(
  n,
  p,
  family = c("gaussian", "binomial"),
  density = 1,
  rho = 0,
  s = 5,
  rho_type = c("constant", "auto"),
  beta_type = 1,
  snr = 2
)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of predictors}

\item{family}{Response type}

\item{density}{Density of design matrix}

\item{rho}{Correlation between predictors (corresponding to choice in \code{rho_type})}

\item{s}{Number of signals (nonzero coefficients)}

\item{rho_type}{Type of correlation between predictors. \code{"constant"} gives uniform
correlation between all predictors. \code{"auto"} means that predictors are
"autocorrelated" with adjacent predictors by the amount specified in \code{rho}.}

\item{beta_type}{Type of coefficients to generate}

\item{snr}{Signal-to-noise ratio}
}
\value{
A list with \code{X}: the design matrix, and \code{y}: the response.
}
\description{
Generate Response and Design Matrix
}
