% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoPath.R
\name{lassoPath}
\alias{lassoPath}
\title{Lasso Path with Hessian Screening Rules}
\usage{
lassoPath(
  X,
  y,
  family = c("gaussian", "binomial"),
  standardize = TRUE,
  screening_type = c("working", "edpp", "gap_safe", "strong"),
  path_length = 100L,
  maxit = 1e+05,
  tol_infeas = 0.001,
  tol_gap = 1e-04,
  gamma = 0.01,
  force_kkt_check = FALSE,
  line_search = TRUE,
  verbosity = 0
)
}
\arguments{
\item{X}{The predictor matrix}

\item{y}{The reponse vector}

\item{family}{The name of the family, "gaussian" or "logistic"}

\item{standardize}{Whether to standardize the predictors}

\item{screening_type}{Screening rule}

\item{path_length}{The (desired) length of the lasso path}

\item{maxit}{Maximum number of iterations for Coordinate Descent loop}

\item{tol_infeas}{Tolerance threshold for maximum infeasibility}

\item{tol_gap}{Tolerance threshold for duality gap}

\item{gamma}{Percent of strong approximation to add to Hessian approximation}

\item{force_kkt_check}{Whether to force KKT checks even when screening rule
is safe}

\item{line_search}{Use line search when \code{family = "binomial"}}

\item{verbosity}{Controls the level of verbosity. 0 = no output.}
}
\description{
Lasso Path with Hessian Screening Rules
}
