% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookAheadScreening-package.R
\docType{package}
\name{LookAheadScreening-package}
\alias{LookAheadScreening}
\alias{LookAheadScreening-package}
\title{LookAheadScreening: Safe Look-Ahead Screening for the Lasso}
\description{
Procedures to fit the lasso using safe look-ahead screening rules.
}
\author{
\strong{Maintainer}: Johan Larsson \email{johan.larsson@stat.lu.se} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Authors:
\itemize{
  \item Jonas Wallin \email{jonas.wallin@stat.lu.se} (\href{https://orcid.org/0000-0003-0381-6593}{ORCID})
}

}
\keyword{internal}
