% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoPath.R
\name{lassoPath}
\alias{lassoPath}
\title{Lasso Path with Hessian Screening Rules}
\usage{
lassoPath(
  X,
  y,
  standardize = TRUE,
  screening_type = c("gap_safe", "gap_safe_lookahead"),
  path_length = 100L,
  maxit = 1e+05,
  tol_infeas = 1e-04,
  tol_gap = 1e-05,
  check_kkt = FALSE,
  verbosity = 0
)
}
\arguments{
\item{X}{The predictor matrix}

\item{y}{The reponse vector}

\item{standardize}{Whether to standardize the predictors}

\item{screening_type}{Screening rule}

\item{path_length}{The (desired) length of the lasso path}

\item{maxit}{Maximum number of iterations for Coordinate Descent loop}

\item{tol_infeas}{Tolerance threshold for maximum infeasibility}

\item{tol_gap}{Tolerance threshold for duality gap}

\item{check_kkt}{Whether to force KKT checks}

\item{verbosity}{Controls the level of verbosity. 0 = no output.}
}
\description{
Lasso Path with Hessian Screening Rules
}
